pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-reader-desktop"
version = "0.1.0"
description = "Desktop app for reading papers, generating answers, notes, and PDF reports"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11",
    "pypdf>=5.0",
    "pytesseract>=0.5",
    "pdf2image>=1.16",
    "python-docx>=1.1",
    "reportlab>=4.2",
    "requests>=2.32",
]

[project.scripts]
paper-reader = "paper_reader.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
